requirements.txt
appdirs==1.4.4
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
distlib==0.3.1
Django==3.1.4
django-crispy-forms==1.10.0
django-widget-tweaks==1.4.8
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
Pillow==8.1.0
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
virtualenv==20.2.2
virtualenvwrapper-win==1.2.6
whitenoise==5.2.0
